+++ src/DetailsView/reports/report-generator-provider.ts
@@ -7,7 +7,7 @@ import { AssessmentReportHtmlGenerator } from './assessment-report-html-generato
 import { ReportGenerator } from './report-generator';
 import { ReportGeneratorV1 } from './report-generator-v1';
 import { ReportGeneratorV2 } from './report-generator-v2';
-import { ReportHtmlGeneratorV1 } from './report-html-generator';
+import { ReportHtmlGeneratorV1 } from './report-html-generator-v1';
 import { ReportNameGenerator } from './report-name-generator';
 
 export type ReportGeneratorProvider = {
