+++ src/DetailsView/reports/report-html-generator-v1.tsx
